move modules to app and split gereral components
