+++ client/src/app/dashboard/main.vue
