+++ client/src/components/root/navbar.vue
