+++ client/src/config.js
@@ -1,3 +1,3 @@
 
-// export const apiUrl = 'https://spa.vedcasts.com.br/api/'
-export const apiUrl = 'http://localhost:8000/api/'
+export const apiUrl = 'https://spa.vedcasts.com.br/api/'
+// export const apiUrl = 'http://localhost:8000/api/'
